Add a testing framework to setup the database and repo
This uses lazy_static to setup a few things before running the tests in
parallel.

One thing to note about this setup, many of the recipe tests are in
test_v0_recipe_etc because that depend on each other and make changes to
the git repo and examine those changes. eg. recipes/list would fail if
it were to run after adding the new recipe.

Since there is no way to control execution order, put them into one
function until there is a better solution.